ista_pv/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
python-dotenv>=1.0
psycopg2-binary>=2.9
django-crispy-forms>=1.14
openpyxl>=3.0
pandas>=2.0
django-fernet-fields>=0.6
django-environ>=0.9
drf-yasg>=1.22
cryptography>=40.0
celery>=5.4
redis>=4.5
